+++ test/test_calculation.py
@@ -1,10 +1,9 @@
-import pandas as pd
-
 from land_consumption.calculation import calculate_land_consumption
 
 
 import geopandas as gpd
-from land_consumption.utils import SQM_TO_HA_FACTOR, LandObjectCategory, LandUseCategory
+from land_consumption.utils import LandObjectCategory, LandUseCategory
+from approvaltests import verify
 
 
 def test_calculate_land_consumption():
@@ -26,22 +25,6 @@ def test_calculate_land_consumption():
         }
     )
 
-    expected_data = pd.DataFrame(
-        {
-            'Land Use Object': ['Buildings', 'Agricultural land', 'Natural land', 'Other'],
-            'Land Use Class': ['Commercial', '', '', 'Other land uses'],
-            'Total Land Area [ha]': [0.4, 0.3, 0.05, 0.05],
-            '% of Consumed Land Area': [100, None, None, None],
-            '% of Settled Land Area': [53.33, 40.0, None, 6.67],
-        }
-    )
-
     result = calculate_land_consumption(area_df)
 
-    assert 'Total Land Area [ha]' in result.columns
-    assert '% of Settled Land Area' in result.columns
-    assert '% of Consumed Land Area' in result.columns
-
-    assert result.loc[0, 'Total Land Area [ha]'] == 4000 * SQM_TO_HA_FACTOR
-
-    pd.testing.assert_frame_equal(result, expected_data)
+    verify(result.to_json(indent=2))